Shuffle LinearAlgebra tests to the front of the test queue
They take much longer than the rest of the tests, so if they're
run at the end, they often determine the length of the entire test
run. Instead, try running them at the start while there's still
plenty of work to be done. Hopefully this will reduce overall duration
of CI tests. If this doesn't work out as planned, we can try a more
fine grained strategy to load balancing (e.g. by checking in a table
of approximate runtimes and sorting according to that).