inference: restore precision lost due to previous bugfix
Since Tuple is widened slowly, we can end up at Any faster than we want
to. Before jumping to Any, first try a very simple Tuple as the Union
element and see if that reduces the overall complexity sufficiently.
This isn't a correctness change, but it should improve the transitive
stability of the tmerge algorithm.